Made a lot of progress
Metroid now atleast draws the titlescreen and the first screen of the game. The titlescreen is really glitchy indicating that we are doing something wrong when switching banks in the mapper. Also imporved how the cpu can log the opcodes it executes. It now stores them to a file instead of just printing them.