added basic controller suppoer
The issue now is that the axis sticks a bit, this is becuse if we go from a input straight to no input theres no frame where the input is below the deadzone (when we are at ~0 no axis event is generated) and because of how the controller caches buttons instead of clearing them every frame, this now means that the buttons are stuck. One way to fix this is for the controller to keep track of when it recieved an axis motion event, then every frame we tell the controller to update, if it hasn gotten a new axis motion event that frame it knows to reset the axis.